Made a non-working test of latest attempt at adding features:
I'm trying to fix the following:

Pull questions from list in a random order weighted toward early questions

Ensure there is a json file of question-answer pairs. I will commit new versions of this file over time in such a way that earlier entries are more recently added questions (I might add timestamp information at some point but simply ordering them might be enough for now).
Every time a user starts the quiz, the webpage will fill a list (actually some appropriate container type like a priority queue or stack... not sure which yet) from the json file, call it the "remaining questions" and proceed to pick the first question.
Every time a question must be picked, start the following picking procedure:
Check if the list of remaining questions is empty and if it is then stop the quiz and go to the results page.
Every time a question must be picked and the list was nonempty, check if the last focussed question was the last in the list of remaining questions and if so then set the first item in the list as focussed, otherwise set the next item from last focussed.
Decide whether or not to select the focussed item in the list based on a random value: if a random value from 0 to 1 is less than some constant threshold then select the focussed item. If the random value is greater than or equal to the threshold then start the picking procedure again.
When an item is selected, remove it from the list, pose it to the user. They must provide an answer and once they submit it, they are taken to a page where they verify if it is correct against a reference answer and once they do so a question is to be picked by the picking procedure again.
On every question page there is also an option to end the quiz presented (the current question is ignored by the scoring)
The scoring results should tally the answered questions to correctly answered questions and also list what those answers were against their questions and their reference answers in a table for review.